generation/qa_engine.py
from typing import List, Dict, Any, Optional, TYPE_CHECKING
import time
import logging

if TYPE_CHECKING:
    from your_vector_store_module import MultiModalVectorStore  # Adjust this import path accordingly

class QAEngine:
    """
    Generates answers using retrieved context from a vector store.

    Attributes:
        vector_store (MultiModalVectorStore): The vector store instance used for document retrieval.

    Methods:
        generate_answer: Retrieves relevant documents and generates an answer with citations.
    """

    def __init__(self, vector_store: 'MultiModalVectorStore'):
        """
        Initialize QAEngine with a vector store.

        Args:
            vector_store (MultiModalVectorStore): Vector store for retrieving relevant contexts.
        """
        self.vector_store = vector_store
        self.logger = logging.getLogger(__name__)

    def generate_answer(self, query: str, use_hybrid: bool = True, top_k: int = 5) -> Dict[str, Any]:
        """
        Generate an answer with citations using the retrieved context.

        Args:
            query (str): The user query string.
            use_hybrid (bool, optional): Whether to use hybrid search or simple retrieval. Defaults to True.
            top_k (int, optional): Number of top documents to retrieve. Defaults to 5.

        Returns:
            Dict[str, Any]: A dictionary including:
                - 'answer': The generated answer string.
                - 'sources': List of source metadata and excerpts.
                - 'context': Concatenated context from retrieved documents.
                - 'retrieval_time_ms': Time spent on retrieval in milliseconds.
                - 'generation_time_ms': Time spent on answer generation in milliseconds.
        """
        start_time = time.time()

        try:
            # Retrieve relevant context documents
            if use_hybrid:
                retrieved = self.vector_store.hybrid_search(query, top_k=top_k)
            else:
                retrieved = self.vector_store.retrieve(query, top_k=top_k)
        except Exception as e:
            self.logger.error(f"Error during document retrieval: {e}")
            retrieved = []

        retrieval_time = time.time() - start_time

        if not retrieved:
            self.logger.warning("No documents retrieved for the query.")
            return {
                'answer': "No relevant documents found to answer your query.",
                'sources': [],
                'context': "",
                'retrieval_time_ms': int(retrieval_time * 1000),
                'generation_time_ms': 0
            }

        context = self._build_context(retrieved)
        sources = self._extract_sources(retrieved)

        generation_start = time.time()
        # Generate answer using LLM (currently placeholder)
        answer = self._generate_with_llm(query, context)
        generation_time = time.time() - generation_start

        return {
            'answer': answer,
            'sources': sources,
            'context': context,
            'retrieval_time_ms': int(retrieval_time * 1000),
            'generation_time_ms': int(generation_time * 1000)
        }

    def _build_context(self, retrieved_docs: List[Dict[str, Any]]) -> str:
        """
        Build a concatenated context string from retrieved documents.

        Args:
            retrieved_docs (List[Dict[str, Any]]): List of retrieved document dictionaries.

        Returns:
            str: Concatenated string of source information and their content snippets.
        """
        context_parts = []
        for i, doc in enumerate(retrieved_docs):
            page = doc.get('metadata', {}).get('page', 'unknown')
            content_snippet = doc.get('content', '')[:500].replace('\n', ' ')
            context_parts.append(f"[Source {i+1}, Page {page}]: {content_snippet}")
        return "\n\n".join(context_parts)

    def _extract_sources(self, retrieved_docs: List[Dict[str, Any]]) -> List[Dict[str, Any]]:
        """
        Extract source metadata and content snippets from retrieved documents.

        Args:
            retrieved_docs (List[Dict[str, Any]]): List of retrieved document dictionaries.

        Returns:
            List[Dict[str, Any]]: List of dictionaries containing source info.
        """
        sources = []
        for doc in retrieved_docs:
            metadata = doc.get('metadata', {})
            source_info = {
                'type': metadata.get('type', 'unknown'),
                'page': metadata.get('page', 'unknown'),
                'confidence': 1 - doc.get('distance', 0),
                'content': (doc.get('content', '')[:200] + '...').replace('\n', ' ')
            }
            sources.append(source_info)
        return sources

    def _generate_with_llm(self, query: str, context: str) -> str:
        """
        Generate an answer using an LLM.

        In production, replace with actual API call, for example:

        ```
        from anthropic import Anthropic
        client = Anthropic(api_key="your-key")

        message = client.messages.create(
            model="claude-sonnet-4-20250514",
            max_tokens=1000,
            messages=[{
                "role": "user",
                "content": f"Based on this context:\n{context}\n\nAnswer: {query}"
            }]
        )
        return message.content[0].text
        ```

        Returns a placeholder response for testing.

        Args:
            query (str): The user query string.
            context (str): Concatenated context from retrieved documents.

        Returns:
            str: Generated answer string.
        """
        return (
            f"Based on the provided documents, here is the answer to "
            f"'{query}'. [This would be generated by an LLM in production]"
        )
